+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - removed container name and storing all containers
 - make_default is renamed to use_container
 - Resolvers are moved to resolvers module
+- added named dependency to @inject
 
 ## 1.1.2